requirements.txt
fastapi~=0.111.0
redis~=5.0.4
SQLAlchemy~=2.0.30
uvicorn~=0.29.0
jsonrpcserver~=5.0.9
PyYAML~=6.0.1